update atom symlinking to work better

+++ dotfiles/atom.rb
@@ -5,7 +5,9 @@ dep 'dotfiles-atom' do
     !raw_shell("find ~/.atom -type l").stdout.empty?
   }
   meet{
-    log_shell "moving ~/.atom → ~/dotfiles/atom", "mv ~/.atom ~/dotfiles/atom"
+    if File.exist?('~/.atom')
+        log_shell "moving ~/.atom → ~/dotfiles/atom", "mv ~/.atom ~/dotfiles/atom"
+    end
     log_shell "symlinking ~/dotfiles/atom → ~/.atom", "ln -fs ~/dotfiles/atom ~/.atom"
   }
 end
